Char, Byte, SByte, CharUnicodeInfo 등

+++ CSClasses/CSClasses/BitConverter.inl
@@ -377,19 +377,19 @@ namespace System
 			}
 		}
 
-		int start = 0;
+		size_t start = 0;
 
 		if (bin.length() / 2 < r.size())
 		{
 			start = r.size() - (bin.length() / 2);
 		}
 
-		for (int i = 0; i < start; i++)
+		for (size_t i = 0; i < start; i++)
 		{
 			r[i] = 0;
 		}
 
-		for (int i = 0; i < bin.length() / 2; i++)
+		for (size_t i = 0; i < bin.length() / 2; i++)
 		{
 			String str = bin.substr(i * 2, 2);
 			Byte b = 0;
